[Tests] Separate and introduce MockCompiler
Let's separate MockCompiler from ToolchainEnv.test and introduce it.

ONE-vscode-DCO-1.0-Signed-off-by: Yongseop Kim <yons.kim@samsung.com>
